requirements.txt
aiogram==3.*
aiohttp
matplotlib
python-dotenv